requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python-headless==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
